requirements.txt
Pillow>=10.0
psutil>=5.9
pyte>=0.8
pywinpty>=2.0
